Update alloca16 macro to cast to int size of pointer.
(cherry picked from commit 6a5b040111f415b9552353f3c88819f9649496d2)

+++ Pulp/Common/Math/XWinding.cpp
@@ -12,7 +12,7 @@ namespace
 	static const int MAX_WORLD_SIZE = (MAX_WORLD_COORD - MIN_WORLD_COORD);
 
 #if 1
-	#define alloca16(numBytes) ((void *)((((int)_alloca( (numBytes)+15 )) + 15) & ~15))
+	#define alloca16(numBytes) ((void *)((((uintptr_t)_alloca( (numBytes)+15 )) + 15) & ~15))
 #else
 	template<typename T>
 	X_INLINE T* Alloca16(size_t num)
